refactor: dedupe CodeFuse model/timestamp helpers and align cost source
- Extract normalizeCodeFuseModel to core (shared by parser and CLI snapshot path)
- Extract parseTimestamp to core, replacing duplicated CLI/parser copies
- Snapshot fallback now derives costSource from resolvePrice() like the parser,
  so a priced model computing zero cost is no longer mislabeled 'unknown'
