+++ packages/core/src/index.ts
@@ -2,6 +2,8 @@ export * from './types.js'
 export * from './provider.js'
 export * from './pricing.js'
 export * from './qoder-model.js'
+export * from './codefuse-model.js'
+export * from './timestamp.js'
 export * from './record-id.js'
 export * from './parsers/index.js'
 export * from './aggregator.js'
